requirements.txt
pymongo==4.13.0
chromadb==1.0.9
qdrant-client
supabase==2.15.1
python-dotenv==1.0.1
sentence_transformers
openai==1.70.0
google-genai
pandas